+++ changelog.md
@@ -2,10 +2,15 @@
 
 ## Next
 
+- Restructured the query input arguments to an object type.
+- Renamed table by scope to `table entries`.
+
 ### Minor
 
 - table_row_arguments scope is now a GraphQLString instead of a name type.
 - Now throws a GraphQL error for mutation adding EOSIO specific error details.
+- Table_row_arguments scope is now a GraphQLString instead of a name type.
+- Now throws a GraphQL error for mutation adding EOSIO specific error details to the return value.
 
 ### Patch